// types/route-names.ts
/**
 * @author Muhammad Omran
 * @date 20-05-2021
 * @description define App route name
 */

/**
 * enum the defines App Route names
 *
 * @enum
 * @exports
 */
export enum Routes {
  SPLASH = "SPLASH",
  TIMER = "TIMER",
  SUCCESS = "SUCCESS",
  HOME_ROUTE = "HOME_ROUTE",
  HOME = "HOME",
  ANNOUNCEMENTS = "ANNOUNCEMENTS",
  ADD_HABIT = "ADD_HABIT",
  IDENTITY_REINFORCEMENT = "IDENTITY_REINFORCEMENT",
  VIEW_HABIT = "VIEW_HABIT",
}